Move StopKeywordInspectionTests for QuickFixes into relevant files under RD.QuickFixes
